+++ HISTORY.md
@@ -2,6 +2,11 @@
 
 ---
 
+## 0.1.3
+
+* `NEW` cellWidth/cellHeight support array type.
+* `CHANGED` matrix height will get the length of the larget vm sub array.
+
 ## 0.1.2
 
 * `FIXED` verify if vm[0] exsits before using